Move entities package to database package
