+++ app/src/main/java/com/shapeapp/shape/data/database/entities/Message.kt
@@ -1,4 +1,4 @@
-package com.shapeapp.shape.data.entities
+package com.shapeapp.shape.data.database.entities
 
 import android.os.Parcel
 import android.os.Parcelable
