+++ app/src/main/java/com/shapeapp/shape/data/network/backendcalls/BackendApi.kt
@@ -1,6 +1,6 @@
 package com.shapeapp.shape.data.network.backendcalls
 
-import com.shapeapp.shape.data.entities.Card
+import com.shapeapp.shape.data.database.entities.Card
 import retrofit2.Call
 import retrofit2.http.GET
