+++ app/src/main/java/com/shapeapp/shape/ui/dialogs/MyPublicSharesBSDViewModel.kt
@@ -2,7 +2,7 @@ package com.shapeapp.shape.ui.dialogs
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
-import com.shapeapp.shape.data.entities.Card
+import com.shapeapp.shape.data.database.entities.Card
 import com.shapeapp.shape.data.repositories.CardsRepository
 
 /**
