+++ app/src/main/java/com/shapeapp/shape/ui/recyclerviews/adapters/MessagesFullRecyclerViewAdapter.kt
@@ -8,7 +8,7 @@ import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.shapeapp.shape.R
-import com.shapeapp.shape.data.entities.Message
+import com.shapeapp.shape.data.database.entities.Message
 import com.shapeapp.shape.internal.mockupsmakers.DrawablesMockups
 
 /**
