+++ app/src/main/java/com/shapeapp/shape/ui/recyclerviews/interfaces/RecyclerViewMessageClickListener.kt
@@ -2,7 +2,7 @@ package com.shapeapp.shape.ui.recyclerviews.interfaces
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
-import com.shapeapp.shape.data.entities.Message
+import com.shapeapp.shape.data.database.entities.Message
 
 /**
  * Interface definition for a callback to be invoked when a message in [RecyclerView] is clicked.
